Model: move activeCustomer from AddressBook to ModelManager
AddressBook stores activeCustomer and has commission methods that assume
activeCustomer's UniqueCustomerList is the target list.

Let's add the customer parameter to AddressBook commission methods and
move activeCustomer to ModelManager to make AddressBook easier to test
in the future.
