+++ src/main/java/seedu/address/model/ReadOnlyAddressBook.java
@@ -1,7 +1,6 @@
 package seedu.address.model;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.commission.Commission;
 import seedu.address.model.customer.Customer;
 
 /**
@@ -14,6 +13,4 @@ public interface ReadOnlyAddressBook {
      * This list will not contain any duplicate customers.
      */
     ObservableList<Customer> getCustomerList();
-
-    ObservableList<Commission> getCommissionList();
 }
