+++ src/test/java/seedu/address/model/AddressBookTest.java
@@ -18,7 +18,6 @@ import org.junit.jupiter.api.Test;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.model.commission.Commission;
 import seedu.address.model.customer.Customer;
 import seedu.address.model.customer.exceptions.DuplicateCustomerException;
 import seedu.address.testutil.CustomerBuilder;
@@ -97,11 +96,6 @@ public class AddressBookTest {
         @Override
         public ObservableList<Customer> getCustomerList() {
             return customers;
-        }
-
-        @Override
-        public ObservableList<Commission> getCommissionList() {
-            throw new AssertionError("This method should not be called.");
         }
     }
